+++ cas-server-core-events-mongo/src/test/java/org/jasig/cas/support/events/mongo/MongoDbCasEventRepositoryTests.java
@@ -1,11 +1,7 @@
 package org.jasig.cas.support.events.mongo;
 
-import org.jasig.cas.mock.MockTicketGrantingTicket;
-import org.jasig.cas.support.events.CasTicketGrantingTicketCreatedEvent;
-import org.jasig.cas.support.events.dao.CasEvent;
+import org.jasig.cas.support.events.AbstractCasEventRepositoryTests;
 import org.jasig.cas.support.events.dao.CasEventRepository;
-import org.jasig.cas.ticket.TicketGrantingTicket;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
@@ -19,22 +15,13 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration("classpath:/mongo-cloudtest-eventscontext.xml")
-public class MongoDbCasEventRepositoryTests {
+public class MongoDbCasEventRepositoryTests extends AbstractCasEventRepositoryTests {
 
     @Autowired
     private CasEventRepository repository;
 
-    @Test
-    public void verifySave() {
-        final TicketGrantingTicket ticket = new MockTicketGrantingTicket("casuser");
-        final CasTicketGrantingTicketCreatedEvent event = new CasTicketGrantingTicketCreatedEvent(this, ticket);
-
-        final CasEvent dto = new CasEvent();
-        dto.setType(event.getClass().getCanonicalName());
-        dto.putTimestamp(event.getTimestamp());
-        dto.putCreationTime(event.getTicketGrantingTicket().getCreationTime());
-        dto.putId(event.getTicketGrantingTicket().getId());
-
-        this.repository.save(dto);
+    @Override
+    public CasEventRepository getRepositoryInstance() {
+        return this.repository;
     }
 }
