+++ cas-server-core-events/build.gradle
@@ -6,6 +6,7 @@ dependencies {
     compile project(':cas-server-core-web')
     compile libraries.spring
     compile libraries.hibernate
+    testCompile project(path: ":cas-server-core-tickets", configuration: "tests")
 }
